tests/wpt/tests/css/css-align/animation/align-no-interpolation.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/w3c/csswg-drafts/issues/4441">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<body>
<script>
test_no_interpolation({
  property: 'align-content',
  from: 'initial',
  to: 'baseline'
});

test_no_interpolation({
  property: 'align-items',
  from: 'initial',
  to: 'baseline'
});

test_no_interpolation({
  property: 'align-self',
  from: 'initial',
  to: 'baseline'
});
</script>
